requirements.txt
Django~=4.1
djangorestframework~=3.13
django-cors-headers~=3.13
psycopg2~=2.9
django-minio-backend~=3.3.0
django-import-export~=2.8
djangorestframework-camel-case~=1.3
django-rest-knox~=4.2
channels~=3.0
channels-redis~=3.4
gunicorn~=20.1
uvicorn[standard]~=0.18
celery~=5.2
redis~=4.3
Pillow~=9.2.0
drf-yasg~=1.20
geoip2~=4.6.0
pyyaml~=6.0
ua-parser~=0.16.1
user-agents~=2.2.0
django-user-agents~=0.4.0
pyotp~=2.7
